LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/tables/table-vspace-hspace-expected.txt

x
x
x

FAIL table vspace hspace (quirks mode) assert_equals: marginTop expected "0px" but got "25px"
